requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aniso8601==9.0.1
    # via graphene
annotated-types==0.6.0
    # via pydantic
asgiref==3.7.2
    # via django
dj-database-url==2.0.0
    # via -r requirements.in
django==4.2.4
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-extensions
    #   django-graphql-jwt
    #   django-polymorphic
    #   django-pydantic-field
    #   graphene-django
django-better-admin-arrayfield==1.4.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-graphql-jwt==0.4.0
    # via -r requirements.in
django-json-widget==1.1.1
    # via -r requirements.in
django-polymorphic==3.1.0
    # via -r requirements.in
django-pydantic-field==0.3.3
    # via -r requirements.in
future==1.0.0
    # via django-json-widget
graphene==3.3
    # via
    #   -r requirements.in
    #   django-graphql-jwt
    #   graphene-django
graphene-django==3.1.3
    # via
    #   -r requirements.in
    #   django-graphql-jwt
graphql-core==3.2.3
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==3.2.0
    # via
    #   graphene
    #   graphene-django
gunicorn==20.1.0
    # via -r requirements.in
promise==2.3
    # via graphene-django
psycopg2-binary==2.9.7
    # via -r requirements.in
pydantic==2.6.4
    # via django-pydantic-field
pydantic-core==2.16.3
    # via pydantic
pyjwt==2.8.0
    # via django-graphql-jwt
python-dotenv==1.0.0
    # via -r requirements.in
six==1.16.0
    # via promise
sqlparse==0.4.4
    # via django
text-unidecode==1.3
    # via graphene-django
typing-extensions==4.7.1
    # via
    #   asgiref
    #   dj-database-url
    #   django-pydantic-field
    #   pydantic
    #   pydantic-core
whitenoise==6.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
